Attempt to parse DoRequest with v3 errors
If you try to use DoRequest to hit v3 endpoints, you run into an issue
because the response json is different than v2. By Default, DoRequest
attempts to parse the response according to the v2 error schema.

Changes:
If the error that gets parsed from the response is real (i.e. it has an
error code that's not 0, the default value), it returns the parsed v2
error. Otherwise, it then attempts to parse it as a v3 error.